+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/nodes/RootNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -391,6 +391,11 @@ public abstract class RootNode extends ExecutableNode {
      * uninitialized copy from each initialized node.
      * </ul>
      *
+     * When this method is invoked during compilation preparation and this root node is associated
+     * with a {@link TruffleLanguage language}, that language is entered, but no language context is
+     * entered. In that case, {@link LanguageReference} can be used, but {@link ContextThreadLocal},
+     * {@link ContextLocal}, and {@link ContextReference} cannot be used.
+     *
      * @return an uninitialized copy of this root node if supported.
      * @throws UnsupportedOperationException if not supported
      * @see #isCloneUninitializedSupported()
@@ -561,9 +566,10 @@ public abstract class RootNode extends ExecutableNode {
      * for safety.
      * </ul>
      *
-     * Note that during the execution of this method, no language context is entered. Therefore, you
-     * cannot use {@link ContextThreadLocal}, {@link ContextLocal}, {@link LanguageReference}, or
-     * {@link ContextReference}.
+     * During the execution of this method, if this root node is associated with a
+     * {@link TruffleLanguage language}, that language is entered, but no language context is
+     * entered. In that case, {@link LanguageReference} can be used, but
+     * {@link ContextThreadLocal}, {@link ContextLocal}, and {@link ContextReference} cannot be used.
      *
      * @param rootCompilation <code>true</code> if this is a root compilation; <code>false</code> if
      *            inlining this root.
@@ -579,6 +585,19 @@ public abstract class RootNode extends ExecutableNode {
      */
     protected boolean prepareForCompilation(boolean rootCompilation, int compilationTier, boolean lastTier) {
         return true;
+    }
+
+    /**
+     * Visits nodes that may be cloned with this root node. The default implementation visits this
+     * root and its ordinary children. Overrides must invoke the superclass implementation before
+     * visiting additional node trees stored outside of ordinary child fields. Additional node trees
+     * must be visited at most once and must not be reachable through ordinary child fields.
+     *
+     * @param visitor the visitor invoked for each node
+     * @since 25.4
+     */
+    protected void visitCloneableNodes(NodeVisitor visitor) {
+        accept(visitor);
     }
 
     /**